requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.10
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==0.2.2
google-generativeai==0.8.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.62.3
httplib2==0.31.0
idna==3.11
importlib_metadata==7.2.1
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
narwhals==2.8.0
numpy==1.26.4
packaging==23.2
pandas==2.3.3
pillow==10.2.0
proto-plus==1.26.1
protobuf==4.25.8
psycopg2-binary==2.9.9
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.1
redis==5.0.1
referencing==0.37.0
requests==2.32.5
rich==13.9.4
rpds-py==0.27.1
rsa==4.9.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.25
streamlit==1.31.0
tenacity==8.5.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
websockets==14.2
zipp==3.23.0
pyjwt
bcrypt==4.1.2
